Rename payment_channel_state_service -> state_service
